bharat-intelligence-engine/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
pydantic-settings==2.5.2
redis==5.1.0
celery==5.4.0
kafka-python==2.0.2
elasticsearch==8.15.1
httpx==0.27.2
beautifulsoup4==4.12.3
scrapy==2.11.2
playwright==1.47.0
tenacity==9.0.0
python-json-logger==2.0.7
pandas==2.2.3
numpy==2.1.1
python-dotenv==1.0.1
scrapy-playwright==0.0.37
